Moved Helper files and fixed imports
