+++ characters/src/helper/getForExpose/getCombatSkills.ts
