+++ characters/src/helper/getForExpose/getCoreAttributes.ts
