+++ characters/src/helper/getForExpose/getEquippedArmor.ts
