+++ characters/src/helper/getForExpose/getEquippedWeapons.ts
