+++ characters/src/helper/getForExpose/getInventory.ts
